PostgreSQL/requirements.txt
psycopg2-binary==2.8.6
SQLAlchemy==1.4.7
